refactor(swiftpm): Move the mapping to ORT's model into SwiftPM
Make `SwiftPmModel` dedicated to just the upstream swift models and
it's parsing, by moving the mapping to ORT's model into `SwiftPM`. This
simplifies refactoring and improves encapsulation as visibility can be
reduced.
